database_and_others/requirements.txt
pinecone-client==3.0.2
boto3==1.34.14
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4